UPDATE ::: Adding state on the insight if it should delete files on drop or not.  This is used for when we preview and make an insight copy
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@33607 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

+++ src/prerna/util/insight/InsightUtility.java
@@ -192,13 +192,15 @@ public class InsightUtility {
 			// since we do not do this in the clear
 			// i will first grab all the files used
 			// then delete them
-			List<FileMeta> fileData = insight.getFilesUsedInInsight();
-			if (fileData != null && !fileData.isEmpty()) {
-				for (int fileIdx = 0; fileIdx < fileData.size(); fileIdx++) {
-					FileMeta file = fileData.get(fileIdx);
-					File f = new File(file.getFileLoc());
-					f.delete();
-					LOGGER.debug("Successfully deleted File used in insight " + file.getFileLoc());
+			if(insight.isDeleteFilesOnDropInsight()) {
+				List<FileMeta> fileData = insight.getFilesUsedInInsight();
+				if (fileData != null && !fileData.isEmpty()) {
+					for (int fileIdx = 0; fileIdx < fileData.size(); fileIdx++) {
+						FileMeta file = fileData.get(fileIdx);
+						File f = new File(file.getFileLoc());
+						f.delete();
+						LOGGER.debug("Successfully deleted File used in insight " + file.getFileLoc());
+					}
 				}
 			}
